hrms-adkfast2/requirements.txt
pymongo
python-dotenv
google-adk
uvicorn
fastapi
